Fix bug where last conditional block of switch would execute as the default

+++ debug.c
@@ -196,7 +196,7 @@ void show_ast(char* name, ast obj) {
   int nb_children = get_nb_children(obj);
   if (nb_children == 0) nb_children = 1; // Account for value of ast nodes with no child
   for (i = 0; i < nb_children + 1; i++) {
-    printf("%s[%d] = %d\n", name, i, heap[obj + i]);
+    printf("%s[%d] = %d\n", name, i, (int) heap[obj + i]);
   }
 }
